Rename lib/syscache module to its path name lib/cache
Previously, the lib/cache path contains "syscache" module with a different
name. It is confusing when import lib/cache path, but directly use the syscache
module. So rename module name to "cache", which is consistent with its located
path name.
